hw16/requirements.txt
biopython==1.79
seaborn==0.11.2
